Refactor Android part
