+++ samples/Playground/Playground.Droid/Playground.Droid.csproj
@@ -88,6 +88,7 @@
     <Compile Include="Views\Collections\GroupedCollectionPageActivity.cs" />
     <Compile Include="Views\Collections\ProductViewHolder.cs" />
     <Compile Include="Converters\BoolToVisibilityConverter.cs" />
+    <Compile Include="Views\Collections\MovieCollectionViewHolder.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
