+++ samples/Playground/Playground.Droid/Views/Collections/CollectionPageActivity.cs
@@ -4,13 +4,9 @@
 using Android.App;
 using Android.OS;
 using Android.Support.V7.Widget;
-using Android.Views;
-using Android.Widget;
-using FFImageLoading;
 using Playground.Models;
 using Playground.ViewModels.Collections;
 using Softeq.XToolkit.Bindings.Droid.Bindable;
-using Softeq.XToolkit.Bindings.Extensions;
 using Softeq.XToolkit.WhiteLabel.Droid;
 
 namespace Playground.Droid.Views.Collections
@@ -35,43 +31,12 @@ namespace Playground.Droid.Views.Collections
             _recyclerView.SetLayoutManager(new LinearLayoutManager(this));
 
             var adapter = new BindableRecyclerViewAdapter<ItemViewModel, MovieCollectionViewHolder>(
-                ViewModel.ItemModels,
-                Resource.Layout.item_movie)
+                ViewModel.ItemModels)
             {
                 ItemClick = ViewModel.SelectItemCommand
             };
 
             _recyclerView.SetAdapter(adapter);
-        }
-    }
-
-    public class MovieCollectionViewHolder : BindableViewHolder<ItemViewModel>
-    {
-        private readonly ImageView _image;
-        private readonly TextView _name;
-
-        public MovieCollectionViewHolder(View view) : base(view)
-        {
-            _image = view.FindViewById<ImageView>(Resource.Id.item_movie_image);
-            _name = view.FindViewById<TextView>(Resource.Id.item_movie_name);
-        }
-
-        public override void DoAttachBindings()
-        {
-            base.DoAttachBindings();
-
-            ImageService.Instance
-                .LoadUrl(ViewModel.IconUrl)
-                .Into(_image);
-
-            this.Bind(() => ViewModel.Title, () => _name.Text);
-        }
-
-        public override void DoDetachBindings()
-        {
-            base.DoDetachBindings();
-
-            _image.SetImageDrawable(null);
         }
     }
 }
